requirements.txt
aiohttp>=3.10.0
google-generativeai==0.7.2
python-dotenv==1.0.1
flask==3.0.3
gunicorn==22.0.0
